% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known.R
\name{add_cog_group}
\alias{add_cog_group}
\title{Add a cognostic group}
\usage{
add_cog_group(name, fields, description = NA, fn, ...)
}
\arguments{
\item{name}{Name of cognostic group}

\item{fields}{\code{data.frame} of \code{'dimension'} and \code{'type'} columns. \code{dplyr::\link[dplyr]{bind_rows}()} of \code{\link{field_info}} outputs for convenience}

\item{description}{Description of cognostic group}

\item{fn}{function to calculate a cognostic group.  May return a named list or a single row tibble.  Each value of the return data should be the output of \code{\link{cog_desc}}}

\item{...}{ignored}
}
\description{
Add a new cognostic to be used when calculating automatic cognostics.
}
\examples{
\dontrun{
add_cog_group(
  "univariate_continuous",
  field_info("x", "continuous"),
  "univariate metrics for continuous data",
  function(x, ...) {
    x_range <- range(x, na.rm = TRUE)
    list(
      min = cog_desc(x_range[1], "minimum of non NA data"),
      max = cog_desc(x_range[2], "maximum of non NA data"),
      mean = cog_desc(mean(x, na.rm = TRUE), "mean of non NA data"),
      median = cog_desc(median(x, na.rm = TRUE), "median of non NA data"),
      var = cog_desc(var(x, na.rm = TRUE), "variance of non NA data")
    )
  }
)
}
}
