% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layer_cogs.R
\name{add_layer_cogs}
\alias{add_layer_cogs}
\title{Add plot layer cognostics}
\usage{
add_layer_cogs(name, description, cog_groups, kind = "ggplot", ...)
}
\arguments{
\item{name}{Name of plot layer. This should match the output of the \code{"name"} values of \code{\link{layer_info}}}

\item{description}{Description of cognostic group}

\item{cog_groups}{A \code{data.frame} (or \code{tibble}) containing the columns: "cog_group", "cols", "name".  "cog_group" column should contain a string value of a known cognostic group.  "cols" should be a single value or vector of column names to use from the data supplied by \code{\link{layer_info}} during calculations. "name" should contain the final storage name of the cognostic group.}

\item{kind}{String value that will match the output of \code{\link{plot_class}} of the desired plot object}

\item{...}{ignored}
}
\description{
Add a new set of cognostic groups for a given plot layer. If the plot layer is found, the corresponding cognostic groups will be calculated.
}
\examples{
\dontrun{
add_layer_cogs(
  "geom_point",
  kind = "ggplot",
  "scatter plot points",
  tibble::tribble(
    ~ cog_group, ~ cols, ~ name,
    "univariate_continuous", "x", "_x",
    "univariate_continuous", "y", "_y",
    "bivariate_continuous", c("x", "y"), "_bivar",
    "scagnostics", c("x", "y"), "_scagnostic",
    "bivariate_counts", c("x", "y"), "_n"
  )
)
}
}
