% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cog.R
\name{auto_cog}
\alias{auto_cog}
\title{Auto cognostic function}
\usage{
auto_cog(.name, ..., .fn_only = FALSE)
}
\arguments{
\item{.name}{name of a known cognostic}

\item{...}{arguments passed onto the found function}

\item{.fn_only}{boolean that determines if the function should be returned}
}
\description{
Calculate an auto cognostic function given a name
}
\examples{
auto_cog("univariate_continuous", iris$Sepal.Length)
fn <- auto_cog("univariate_continuous", .fn_only = TRUE)
fn(iris$Sepal.Length)
}
