% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocogs-package.R
\docType{package}
\name{autocogs-package}
\alias{autocogs}
\alias{autocogs-package}
\title{autocogs: Automatic Cognostic Summaries}
\description{
Automatically calculates cognostic groups for plot objects and list column plot objects. Results are returned in a nested data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schloerke/autocogs}
  \item Report bugs at \url{https://github.com/schloerke/autocogs/issues}
}

}
\author{
\strong{Maintainer}: Ryan Hafen \email{rhafen@gmail.com} (\href{https://orcid.org/0000-0002-5516-8367}{ORCID}) [thesis advisor]

Authors:
\itemize{
  \item Barret Schloerke \email{schloerke@gmail.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
}

}
\keyword{internal}
