% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known.R
\name{cog}
\alias{cog}
\title{Cognostic}
\usage{
cog(x, desc = NULL)
}
\arguments{
\item{x}{univariate scalar}

\item{desc}{description of \code{x}}
}
\description{
Add a description to a cognostic (subset metric)
}
\examples{
cog(mean(1:10), "mean of 10 numbers")
}
