% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_desc.R
\name{cog_desc}
\alias{cog_desc}
\title{Cognostic}
\usage{
cog_desc(x, desc = NULL)
}
\arguments{
\item{x}{univariate scalar}

\item{desc}{description of \code{x}}
}
\description{
Add a description to a cognostic (subset metric)
}
\examples{
cog_desc(mean(1:10), "mean of 10 numbers")
}
