% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layer_cogs_.R
\name{cog_group}
\alias{cog_group}
\title{Cog group data frame}
\usage{
cog_group(...)
}
\arguments{
\item{...}{sets of three values to fill in 'cog_group', 'cols', and 'name'}
}
\description{
Make a cog group data frame to be passed into \code{\link[=add_layer_cogs]{add_layer_cogs()}}
}
\examples{
cog_group(
  "univariate_discrete", "x", "_x",
  "univariate_counts", "x", "_n"
)
cog_group(
  "univariate_continuous", "x", "_x",
  "univariate_continuous", "y", "_y",
  "bivariate_continuous", c("x", "y"), "_bivar",
  "scagnostics", c("x", "y"), "_scagnostic",
  "bivariate_counts", c("x", "y"), "_n"
)
}
