% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{get_data_list}
\alias{get_data_list}
\title{Data List}
\usage{
get_data_list(p, ...)
}
\arguments{
\item{p}{plot object}
}
\description{
Data List
}
\examples{
require(ggplot2)
p <- ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point(data = mpg, mapping = aes(cty, hwy))
get_data_list(p)
}
