% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_cog_groups.R
\name{known_cog_groups}
\alias{known_cog_groups}
\title{Cognostic Group information}
\usage{
known_cog_groups()
}
\description{
To add more cognostic groups, please see \code{\link{add_cog_group}()}
}
\examples{
known_cog_groups()
}
