% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_layer_cogs.R
\name{known_layer_cogs}
\alias{known_layer_cogs}
\title{Layer Cognostic groups}
\usage{
known_layer_cogs()
}
\description{
Display all layer cognostic information to be paired with information from \code{\link{known_cog_groups}()}.
}
\examples{
known_layer_cogs()
}
