% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cog.R
\name{panel_cogs}
\alias{panel_cogs}
\alias{add_panel_cogs}
\title{Panel cognostics}
\usage{
panel_cogs(dt, panel_col = "panel", ...)

add_panel_cogs(dt, panel_col = "panel", ...)
}
\arguments{
\item{dt}{data to be used}

\item{panel_col}{panel column to be used in \code{dt}}
}
\description{
Return or concatinate panel cognostics.  For each panel (plot) in the panel column, cognostics will be calculated for each panel. The result will be returned in a nested \code{\link[tibble]{tibble}}.
}
