% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class.R
\name{plot_class}
\alias{plot_class}
\alias{plot_class.default}
\alias{plot_class.gg}
\alias{plot_class.ggplot}
\title{Plot class}
\usage{
plot_class(p)

\method{plot_class}{default}(p)

\method{plot_class}{gg}(p)

\method{plot_class}{ggplot}(p)
}
\arguments{
\item{p}{plot object to retrieve class from}
}
\description{
First class of the plot object. Exception is ggplot2 as many objects are of class 'gg'
}
\examples{
library(ggplot2)
p <- qplot(Sepal.Length, Sepal.Width, data = iris)
plot_class(p)
}
