% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_info.R
\name{field_info}
\alias{field_info}
\title{Field Type Information}
\usage{
field_info(
  dimension = c("x", "y", "z", "group", "any"),
  type = c("continuous", "discrete", "date", "any")
)
}
\arguments{
\item{dimension}{field name. Use one of the listed options provided}

\item{type}{field type. Use one of the listed options provided}
}
\description{
Field Type Information
}
