% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocog.R
\name{panel_cogs}
\alias{panel_cogs}
\alias{add_panel_cogs}
\title{Panel cognostics}
\usage{
panel_cogs(dt, panel_col = "panel", ...)

add_panel_cogs(dt, panel_col = "panel", ...)
}
\arguments{
\item{dt}{data to be used}

\item{panel_col}{panel column to be used in \code{dt}}

\item{...}{parameters passed to \code{\link[=layer_info]{layer_info()}}}
}
\description{
Return or concatenate panel cognostics.  For each panel (plot) in the panel column, cognostics will be calculated for each panel. The result will be returned in a nested \code{\link[tibble:tibble]{tibble::tibble()}}.
}
